+++ src/internal/m365/collection/exchange/mail_container_cache.go
@@ -2,11 +2,13 @@ package exchange
 
 import (
 	"context"
+	"time"
 
 	"github.com/alcionai/clues"
 	"github.com/microsoftgraph/msgraph-sdk-go/models"
 
 	"github.com/alcionai/corso/src/pkg/fault"
+	"github.com/alcionai/corso/src/pkg/logger"
 	"github.com/alcionai/corso/src/pkg/path"
 	"github.com/alcionai/corso/src/pkg/services/m365/api"
 	"github.com/alcionai/corso/src/pkg/services/m365/api/graph"
@@ -97,6 +99,10 @@ func (mc *mailContainerCache) Populate(
 	baseID string,
 	baseContainerPath ...string,
 ) error {
+	start := time.Now()
+
+	logger.Ctx(ctx).Info("populating container cache")
+
 	if err := mc.init(ctx); err != nil {
 		return clues.Wrap(err, "initializing")
 	}
@@ -108,8 +114,10 @@ func (mc *mailContainerCache) Populate(
 		mc.userID,
 		"",
 		false)
+	ctx = clues.Add(ctx, "num_enumerated_containers", len(containers))
+
 	if err != nil {
-		return clues.Wrap(err, "enumerating containers")
+		return clues.WrapWC(ctx, err, "enumerating containers")
 	}
 
 	for _, c := range containers {
@@ -130,6 +138,10 @@ func (mc *mailContainerCache) Populate(
 	if err := mc.populatePaths(ctx, errs); err != nil {
 		return clues.Wrap(err, "populating paths")
 	}
+
+	logger.Ctx(ctx).Infow(
+		"done populating container cache",
+		"duration", time.Since(start))
 
 	return el.Failure()
 }