+++ src/internal/m365/service/sharepoint/export_test.go
@@ -60,49 +60,108 @@ func (suite *ExportUnitSuite) TestExportRestoreCollections() {
 	defer flush()
 
 	var (
-		driveID       = "driveID1"
-		driveName     = "driveName1"
-		itemName      = "name1"
-		exportCfg     = control.ExportConfig{}
-		dpb           = odConsts.DriveFolderPrefixBuilder(driveID)
-		expectedPath  = path.LibrariesCategory.HumanString() + "/" + driveName
-		expectedItems = []export.Item{
-			{
-				ID:   "id1.data",
-				Name: itemName,
-				Body: io.NopCloser((bytes.NewBufferString("body1"))),
-			},
-		}
+		driveID   = "driveID1"
+		driveName = "driveName1"
+		exportCfg = control.ExportConfig{}
+		dpb       = odConsts.DriveFolderPrefixBuilder(driveID)
 	)
 
-	p, err := dpb.ToDataLayerSharePointPath("t", "u", path.LibrariesCategory, false)
-	assert.NoError(t, err, "build path")
-
 	table := []struct {
-		name     string
-		itemInfo details.ItemInfo
+		name          string
+		itemName      string
+		itemID        string
+		itemInfo      details.ItemInfo
+		getCollPath   func(t *testing.T) path.Path
+		statsCat      path.CategoryType
+		expectedItems []export.Item
+		expectedPath  string
 	}{
 		{
-			name: "OneDriveLegacyItemInfo",
+			name:     "OneDriveLegacyItemInfo",
+			itemName: "name1",
+			itemID:   "id1.data",
 			itemInfo: details.ItemInfo{
 				OneDrive: &details.OneDriveInfo{
 					ItemType:  details.OneDriveItem,
-					ItemName:  itemName,
+					ItemName:  "name1",
 					Size:      1,
 					DriveName: driveName,
 					DriveID:   driveID,
 				},
 			},
+			getCollPath: func(t *testing.T) path.Path {
+				p, err := dpb.ToDataLayerSharePointPath("t", "u", path.LibrariesCategory, false)
+				assert.NoError(t, err, "build path")
+
+				return p
+			},
+			statsCat:     path.FilesCategory,
+			expectedPath: path.LibrariesCategory.HumanString() + "/" + driveName,
+			expectedItems: []export.Item{
+				{
+					ID:   "id1.data",
+					Name: "name1",
+					Body: io.NopCloser((bytes.NewBufferString("body1"))),
+				},
+			},
 		},
 		{
-			name: "SharePointItemInfo",
+			name:     "SharePointItemInfo, Libraries Category",
+			itemName: "name1",
+			itemID:   "id1.data",
 			itemInfo: details.ItemInfo{
 				SharePoint: &details.SharePointInfo{
 					ItemType:  details.SharePointLibrary,
-					ItemName:  itemName,
+					ItemName:  "name1",
 					Size:      1,
 					DriveName: driveName,
 					DriveID:   driveID,
+				},
+			},
+			getCollPath: func(t *testing.T) path.Path {
+				p, err := dpb.ToDataLayerSharePointPath("t", "u", path.LibrariesCategory, false)
+				assert.NoError(t, err, "build path")
+
+				return p
+			},
+			statsCat:     path.FilesCategory,
+			expectedPath: path.LibrariesCategory.HumanString() + "/" + driveName,
+			expectedItems: []export.Item{
+				{
+					ID:   "id1.data",
+					Name: "name1",
+					Body: io.NopCloser((bytes.NewBufferString("body1"))),
+				},
+			},
+		},
+		{
+			name:     "SharePointItemInfo, Lists Category",
+			itemName: "list1",
+			itemID:   "listid1",
+			itemInfo: details.ItemInfo{
+				SharePoint: &details.SharePointInfo{
+					ItemType: details.SharePointList,
+					List: &details.ListInfo{
+						Name:      "list1",
+						ItemCount: 10,
+					},
+				},
+			},
+			getCollPath: func(t *testing.T) path.Path {
+				p, err := path.Elements{"listid1"}.
+					Builder().
+					ToDataLayerSharePointListPath("t", "u", path.ListsCategory, false)
+				assert.NoError(t, err, "build path")
+
+				return p
+			},
+			statsCat:     path.ListsCategory,
+			expectedPath: path.ListsCategory.HumanString() + "/listid1",
+			expectedItems: []export.Item{
+				{
+					ID:   "listid1",
+					Name: "listid1.json",
+					Body: io.NopCloser((bytes.NewBufferString("body1"))),
 				},
 			},
 		},
@@ -115,34 +174,34 @@ func (suite *ExportUnitSuite) TestExportRestoreCollections() {
 			dcs := []data.RestoreCollection{
 				data.FetchRestoreCollection{
 					Collection: dataMock.Collection{
-						Path: p,
+						Path: test.getCollPath(t),
 						ItemData: []data.Item{
 							&dataMock.Item{
-								ItemID: "id1.data",
+								ItemID: test.itemID,
 								Reader: io.NopCloser(bytes.NewBufferString("body1")),
 							},
 						},
 					},
-					FetchItemByNamer: finD{id: "id1.meta", name: itemName},
+					FetchItemByNamer: finD{id: "id1.meta", name: test.itemName},
 				},
 			}
 
 			handler := NewSharePointHandler(api.Client{}, nil)
 			handler.CacheItemInfo(test.itemInfo)
 
-			stats := metrics.ExportStats{}
+			stats := metrics.NewExportStats()
 
 			ecs, err := handler.ProduceExportCollections(
 				ctx,
 				int(version.Backup),
 				exportCfg,
 				dcs,
-				&stats,
+				stats,
 				fault.New(true))
 			assert.NoError(t, err, "export collections error")
 			assert.Len(t, ecs, 1, "num of collections")
 
-			assert.Equal(t, expectedPath, ecs[0].BasePath(), "base dir")
+			assert.Equal(t, test.expectedPath, ecs[0].BasePath(), "base dir")
 
 			fitems := []export.Item{}
 			size := 0
@@ -159,12 +218,12 @@ func (suite *ExportUnitSuite) TestExportRestoreCollections() {
 				fitems = append(fitems, item)
 			}
 
-			assert.Equal(t, expectedItems, fitems, "items")
+			assert.Equal(t, test.expectedItems, fitems, "items")
 
-			expectedStats := metrics.ExportStats{}
-			expectedStats.UpdateBytes(path.FilesCategory, int64(size))
-			expectedStats.UpdateResourceCount(path.FilesCategory)
-			assert.Equal(t, expectedStats, stats, "stats")
+			expectedStats := metrics.NewExportStats()
+			expectedStats.UpdateBytes(test.statsCat, int64(size))
+			expectedStats.UpdateResourceCount(test.statsCat)
+			assert.Equal(t, expectedStats.GetStats(), stats.GetStats(), "stats")
 		})
 	}
 }