+++ src/pkg/services/m365/api/lists_test.go
@@ -332,7 +332,7 @@ func (suite *ListsUnitSuite) TestColumnDefinitionable_LegacyColumns() {
 		name             string
 		getList          func() *models.List
 		length           int
-		expectedColNames map[string]struct{}
+		expectedColNames map[string]any
 	}{
 		{
 			name: "all legacy columns",
@@ -346,7 +346,7 @@ func (suite *ListsUnitSuite) TestColumnDefinitionable_LegacyColumns() {
 				return lst
 			},
 			length:           0,
-			expectedColNames: map[string]struct{}{TitleColumnName: {}},
+			expectedColNames: map[string]any{TitleColumnName: nil},
 		},
 		{
 			name: "title and legacy columns",
@@ -361,7 +361,7 @@ func (suite *ListsUnitSuite) TestColumnDefinitionable_LegacyColumns() {
 				return lst
 			},
 			length:           0,
-			expectedColNames: map[string]struct{}{TitleColumnName: {}},
+			expectedColNames: map[string]any{TitleColumnName: nil},
 		},
 		{
 			name: "readonly and legacy columns",
@@ -376,7 +376,7 @@ func (suite *ListsUnitSuite) TestColumnDefinitionable_LegacyColumns() {
 				return lst
 			},
 			length:           0,
-			expectedColNames: map[string]struct{}{TitleColumnName: {}},
+			expectedColNames: map[string]any{TitleColumnName: nil},
 		},
 		{
 			name: "legacy and a text column",
@@ -391,9 +391,9 @@ func (suite *ListsUnitSuite) TestColumnDefinitionable_LegacyColumns() {
 				return lst
 			},
 			length: 1,
-			expectedColNames: map[string]struct{}{
-				TitleColumnName: {},
-				textColumnName:  {},
+			expectedColNames: map[string]any{
+				TitleColumnName: nil,
+				textColumnName:  nil,
 			},
 		},
 	}
@@ -432,7 +432,7 @@ func (suite *ListsUnitSuite) TestFieldValueSetable() {
 	origFs := models.NewFieldValueSet()
 	origFs.SetAdditionalData(additionalData)
 
-	colNames := map[string]struct{}{}
+	colNames := map[string]any{}
 
 	fs := retrieveFieldData(origFs, colNames)
 	fsAdditionalData := fs.GetAdditionalData()
@@ -509,8 +509,8 @@ func (suite *ListsUnitSuite) TestFieldValueSetable_Location() {
 	origFs := models.NewFieldValueSet()
 	origFs.SetAdditionalData(additionalData)
 
-	colNames := map[string]struct{}{
-		"MyAddress": {},
+	colNames := map[string]any{
+		"MyAddress": nil,
 	}
 
 	fs := retrieveFieldData(origFs, colNames)